fix: remove @ts-nocheck, fix broken CanvasArea import, add type declarations
- Remove @ts-nocheck from SmartPlaybook.tsx and CanvasArea.tsx
- Fix CanvasArea.tsx broken colon-path import (../:components:SmartPlaybook:Field.js)
- Consolidate Field.js: replace 16-line stub with full 530-line canvas implementation
- Delete corrupted :components:SmartPlaybook:Field.js duplicate
- Add forwardRef support to Field.js for canvas ref sharing
- Add typed interfaces to SmartPlaybook.tsx (PlaybookPlayer, PlaybookRoute, etc.)
- Add type annotations to all callback parameters (63 implicit-any fixes)
- Fix dead-code bug: mode === 'delete' check was inside unreachable branch
- Add .d.ts type declarations for all 11 untyped .js components
- Result: npm run build succeeds, 0 TS errors project-wide
