+++ src/components/SmartPlaybook/SmartPlaybook.tsx
@@ -1,4 +1,3 @@
-// @ts-nocheck
 /**
  * SmartPlaybook.js – Main application component
  * - Orchestrates all Smart Playbook functionality
@@ -7,7 +6,6 @@
  */
 
 import React, { useState, useCallback, useEffect, useRef } from 'react';
-import Field from './Field';
 import DebugPanel from './DebugPanel';
 import PlayLibrary from './PlayLibrary';
 import CanvasArea from './components/CanvasArea';
@@ -43,6 +41,56 @@ import Notification from './components/Notification';
 import Onboarding from './components/Onboarding';
 import { AIProvider } from '../../ai-brain/AIContext';
 
+// Internal state types for SmartPlaybook UI layer
+interface PlaybookPlayer {
+  id: string;
+  x: number;
+  y: number;
+  position: string;
+  number: number;
+  selected?: boolean;
+}
+
+interface PlaybookRoute {
+  id: string;
+  playerId: string;
+  points: Array<{ x: number; y: number }>;
+  type: string;
+  color?: string;
+}
+
+interface PlaybookSnapshot {
+  players: PlaybookPlayer[];
+  routes: PlaybookRoute[];
+  action: string;
+  timestamp: number;
+}
+
+interface SavedPlay {
+  id: string;
+  name: string;
+  phase: string;
+  type: string;
+  players: PlaybookPlayer[];
+  routes: PlaybookRoute[];
+}
+
+interface PlaybookNotification {
+  id: number;
+  type: 'success' | 'error' | 'warning';
+  message: string;
+  duration: number;
+}
+
+interface DebugResult {
+  name: string;
+  category: string;
+  passed: boolean;
+  duration: number;
+}
+
+type PlaybookMode = 'view' | 'player' | 'route' | 'delete';
+
 // Constants
 const FIELD_DIMENSIONS = {
   width: 600,
@@ -53,32 +101,32 @@ const TOUCH_THRESHOLD = 20; // pixels for touch detection
 
 const SmartPlaybook = () => {
   // Core state
-  const [players, setPlayers] = useState([]);
-  const [routes, setRoutes] = useState([]);
-  const [selectedPlayerId, setSelectedPlayerId] = useState(null);
-  const [selectedRouteId, setSelectedRouteId] = useState(null);
-  const [mode, setMode] = useState('view'); // view, player, route, delete
-  const [savedPlays, setSavedPlays] = useState([]);
+  const [players, setPlayers] = useState<PlaybookPlayer[]>([]);
+  const [routes, setRoutes] = useState<PlaybookRoute[]>([]);
+  const [selectedPlayerId, setSelectedPlayerId] = useState<string | null>(null);
+  const [selectedRouteId, setSelectedRouteId] = useState<string | null>(null);
+  const [mode, setMode] = useState<PlaybookMode>('view');
+  const [savedPlays, setSavedPlays] = useState<SavedPlay[]>([]);
   const [currentPlayName, setCurrentPlayName] = useState('');
   const [currentPlayPhase, setCurrentPlayPhase] = useState('offense');
   const [currentPlayType, setCurrentPlayType] = useState('pass');
 
   // Route drawing state
   const [isDrawingRoute, setIsDrawingRoute] = useState(false);
-  const [routePoints, setRoutePoints] = useState([]);
+  const [routePoints, setRoutePoints] = useState<Array<{ x: number; y: number }>>([]);
   const [routeType, setRouteType] = useState('custom');
   const [routeColor, setRouteColor] = useState('#ef4444');
 
   // Undo/Redo state
-  const [undoStack, setUndoStack] = useState([]);
-  const [redoStack, setRedoStack] = useState([]);
+  const [undoStack, setUndoStack] = useState<PlaybookSnapshot[]>([]);
+  const [redoStack, setRedoStack] = useState<PlaybookSnapshot[]>([]);
 
   // UI state
   const [showLibrary, setShowLibrary] = useState(false);
   const [showSaveDialog, setShowSaveDialog] = useState(false);
   const [debugMode, setDebugMode] = useState(false);
-  const [debugResults, setDebugResults] = useState([]);
-  const [notifications, setNotifications] = useState([]);
+  const [debugResults, setDebugResults] = useState<DebugResult[]>([]);
+  const [notifications, setNotifications] = useState<PlaybookNotification[]>([]);
   const [showOnboarding, setShowOnboarding] = useState(false);
 
   // Refs
@@ -123,7 +171,7 @@ const SmartPlaybook = () => {
 
   // Prevent double-tap zoom on mobile
   useEffect(() => {
-    const preventDoubleTapZoom = (event) => {
+    const preventDoubleTapZoom = (event: TouchEvent) => {
       if (event.touches.length > 1) {
         event.preventDefault();
       }
@@ -163,7 +211,7 @@ const SmartPlaybook = () => {
   }, [savedPlays]);
 
   // Save current state to undo stack
-  const saveToUndoStack = useCallback((action) => {
+  const saveToUndoStack = useCallback((action: string) => {
     const currentState = {
       players: [...players],
       routes: [...routes],
@@ -207,7 +255,7 @@ const SmartPlaybook = () => {
   }, [undoStack, redoStack, players, routes]);
 
   // Get canvas coordinates from event
-  const getCanvasCoordinates = useCallback((event) => {
+  const getCanvasCoordinates = useCallback((event: React.MouseEvent | React.TouchEvent) => {
     const canvas = canvasRef.current;
     if (!canvas) return null;
 
@@ -215,14 +263,16 @@ const SmartPlaybook = () => {
     const scaleX = canvas.width / rect.width;
     const scaleY = canvas.height / rect.height;
 
-    let clientX, clientY;
-    
-    if (event.touches && event.touches[0]) {
+    let clientX: number, clientY: number;
+
+    if ('touches' in event && event.touches[0]) {
       clientX = event.touches[0].clientX;
       clientY = event.touches[0].clientY;
-    } else {
+    } else if ('clientX' in event) {
       clientX = event.clientX;
       clientY = event.clientY;
+    } else {
+      return null;
     }
 
     return {
@@ -232,7 +282,7 @@ const SmartPlaybook = () => {
   }, []);
 
   // Handle canvas click/touch
-  const handleCanvasEvent = useCallback((event) => {
+  const handleCanvasEvent = useCallback((event: React.MouseEvent | React.TouchEvent) => {
     event.preventDefault();
 
     if (!canvasRef.current) return;
@@ -247,9 +297,9 @@ const SmartPlaybook = () => {
     }
 
     // Handle player selection
-    if (mode === 'view' || mode === 'player') {
-      const clickedPlayer = findPlayerAtPosition(players, coords.x, coords.y, TOUCH_THRESHOLD);
-      
+    if (mode === 'view' || mode === 'player' || mode === 'delete') {
+      const clickedPlayer = findPlayerAtPosition(players, coords.x, coords.y, TOUCH_THRESHOLD) as PlaybookPlayer | null;
+
       if (clickedPlayer) {
         if (mode === 'delete') {
           // Delete player
@@ -278,7 +328,7 @@ const SmartPlaybook = () => {
   }, [mode, isDrawingRoute, players, getCanvasCoordinates, saveToUndoStack]);
 
   // Handle player drag — FIELD_DIMENSIONS is a module constant, no need in deps
-  const handlePlayerDrag = useCallback((playerId, newX, newY) => {
+  const handlePlayerDrag = useCallback((playerId: string, newX: number, newY: number) => {
     // Constrain to field boundaries
     const constrainedX = Math.max(20, Math.min(FIELD_DIMENSIONS.width - 20, newX));
     const constrainedY = Math.max(20, Math.min(FIELD_DIMENSIONS.height - 20, newY));
@@ -287,30 +337,30 @@ const SmartPlaybook = () => {
   }, []);
 
   // Handle player drag end (save to undo stack)
-  const handlePlayerDragEnd = useCallback((playerId) => {
+  const handlePlayerDragEnd = useCallback((playerId: string) => {
     saveToUndoStack('move_player');
     addNotification('success', 'Player moved successfully');
   }, [saveToUndoStack]);
 
   // Add notification helper
-  const addNotification = useCallback((type, message, duration = 3000) => {
+  const addNotification = useCallback((type: 'success' | 'error' | 'warning', message: string, duration = 3000) => {
     const id = Date.now() + Math.random();
     setNotifications(prev => [...prev, { id, type, message, duration }]);
   }, []);
 
   // Remove notification helper
-  const removeNotification = useCallback((id) => {
+  const removeNotification = useCallback((id: number) => {
     setNotifications(prev => prev.filter(n => n.id !== id));
   }, []);
 
   // Handle route selection
-  const handleRouteSelect = useCallback((routeId) => {
+  const handleRouteSelect = useCallback((routeId: string) => {
     setSelectedRouteId(routeId);
     setSelectedPlayerId(null); // Deselect player when selecting route
   }, []);
 
   // Handle route update
-  const handleRouteUpdate = useCallback((routeId, updates) => {
+  const handleRouteUpdate = useCallback((routeId: string, updates: Partial<PlaybookRoute>) => {
     saveToUndoStack('update_route');
     setRoutes(prev => prev.map(route => 
       route.id === routeId ? { ...route, ...updates } : route
@@ -318,14 +368,14 @@ const SmartPlaybook = () => {
   }, [saveToUndoStack]);
 
   // Handle route deletion
-  const handleRouteDelete = useCallback((routeId) => {
+  const handleRouteDelete = useCallback((routeId: string) => {
     saveToUndoStack('delete_route');
     setRoutes(prev => removeRoute(prev, routeId));
     setSelectedRouteId(null);
   }, [saveToUndoStack]);
 
   // Handle preset route application
-  const handleApplyPreset = useCallback((routeId, preset) => {
+  const handleApplyPreset = useCallback((routeId: string, preset: { id: string; points: Array<{ x: number; y: number }> }) => {
     saveToUndoStack('apply_preset_route');
     const route = routes.find(r => r.id === routeId);
     if (!route) return;
@@ -345,7 +395,7 @@ const SmartPlaybook = () => {
   }, [routes, players, saveToUndoStack]);
 
   // Start route drawing
-  const startRouteDrawing = useCallback((playerId) => {
+  const startRouteDrawing = useCallback((playerId: string) => {
     const player = players.find(p => p.id === playerId);
     if (!player) return;
 
@@ -358,7 +408,7 @@ const SmartPlaybook = () => {
 
   // Finish route drawing
   const finishRouteDrawing = useCallback(() => {
-    if (routePoints.length < 2) {
+    if (routePoints.length < 2 || !selectedPlayerId) {
       setIsDrawingRoute(false);
       setRoutePoints([]);
       return;
@@ -381,7 +431,7 @@ const SmartPlaybook = () => {
   }, []);
 
   // Load formation
-  const loadFormation = useCallback((formationType) => {
+  const loadFormation = useCallback((formationType: string) => {
     const centerX = FIELD_DIMENSIONS.width / 2;
     const centerY = FIELD_DIMENSIONS.height / 2;
     
@@ -404,7 +454,7 @@ const SmartPlaybook = () => {
   }, [saveToUndoStack]);
 
   // Save current play
-  const saveCurrentPlay = useCallback((name, phase, type) => {
+  const saveCurrentPlay = useCallback((name: string, phase: string, type: string) => {
     if (!name.trim() || players.length === 0) {
       addNotification('error', 'Please enter a play name and add at least one player.');
       return;
@@ -416,7 +466,7 @@ const SmartPlaybook = () => {
       type: type || currentPlayType,
       players,
       routes
-    });
+    }) as SavedPlay;
 
     setSavedPlays(prev => [...prev, play]);
     setShowSaveDialog(false);
@@ -425,7 +475,7 @@ const SmartPlaybook = () => {
   }, [players, routes, currentPlayPhase, currentPlayType, addNotification]);
 
   // Load play
-  const loadPlay = useCallback((play) => {
+  const loadPlay = useCallback((play: SavedPlay) => {
     saveToUndoStack('load_play');
     setPlayers([...play.players]);
     setRoutes([...play.routes]);
@@ -438,7 +488,7 @@ const SmartPlaybook = () => {
   }, [saveToUndoStack, addNotification]);
 
   // Delete play
-  const deletePlay = useCallback((playId) => {
+  const deletePlay = useCallback((playId: string) => {
     if (window.confirm('Are you sure you want to delete this play?')) {
       setSavedPlays(prev => prev.filter(play => play.id !== playId));
       addNotification('success', 'Play deleted successfully');
@@ -542,7 +592,7 @@ const SmartPlaybook = () => {
             <PlayerControls
               selectedPlayer={players.find(p => p.id === selectedPlayerId)}
               players={players}
-              onUpdatePlayer={(updates) => {
+              onUpdatePlayer={(updates: Partial<PlaybookPlayer>) => {
                 saveToUndoStack('update_player');
                 setPlayers(prev => prev.map(p => 
                   p.id === selectedPlayerId ? { ...p, ...updates } : p
