+++ src/components/SmartPlaybook/components/FormationTemplates.d.ts
@@ -1,0 +1,8 @@
+import React from 'react';
+
+interface FormationTemplatesProps {
+  onLoadFormation: (formationId: string) => void;
+}
+
+declare const FormationTemplates: React.FC<FormationTemplatesProps>;
+export default FormationTemplates;
